+++ Aufgabe2/Buchung/Einnahme.java
@@ -5,7 +5,7 @@ import java.util.Date;
 abstract public class Einnahme
     extends AbstractBuchung
 {
-	public Einnahme(int wert, Date datum) {
+	public Einnahme(float wert, Date datum) {
 		super(wert, datum);
 		assert wert >= 0 : "Einnahmen müssen positiv sein";
 	}
