+++ Aufgabe2/Buchung/Gage.java
@@ -3,7 +3,7 @@ package Buchung;
 import java.util.Date;
 
 public class Gage extends Einnahme {
-	public Gage(int wert, Date datum) {
+	public Gage(float wert, Date datum) {
 		super(wert, datum);
 	}
 }
