+++ Aufgabe2/Buchung/Kosten.java
@@ -5,7 +5,7 @@ import java.util.Date;
 abstract public class Kosten
     extends AbstractBuchung
 {
-	protected Kosten(int wert, Date datum) {
+	protected Kosten(float wert, Date datum) {
 		super(-wert, datum);
 		assert wert >= 0 : "Kosten müssen positiv sein";
 	}
