+++ Aufgabe2/Buchung/RaumMiete.java
@@ -3,7 +3,7 @@ package Buchung;
 import java.util.Date;
 
 public class RaumMiete extends Kosten {
-	public RaumMiete(int wert, Date datum) {
+	public RaumMiete(float wert, Date datum) {
 		super(wert, datum);
 	}
 }
